requirements.txt
wp-engine-api-python
python-dotenv
click
pandas
